version bump

+++ Quilt/src/main/java/software/bernie/geckolib3q/util/GeckoLibUtil.java
@@ -83,7 +83,7 @@ public class GeckoLibUtil {
 	}
 
 	public static AnimationController getControllerForID(AnimationFactory factory, Integer id, String controllerName) {
-		return factory.getOrCreateAnimationData(id).getAnimationControllers().get(controllerName);
+		return factory.getOrCreateAnimationData(id.intValue()).getAnimationControllers().get(controllerName);
 	}
 
 	/**
